remove some rather odd-looking lavp-specific code in LAVPvideo.m
A comment that used to be near this code said "FIXME: for normal operation, first pict in the pictq should have earlier time stamp before currrent time stamp. It is NOT for now.", which suggests this was all about skipping that pict.  But frame_queue_peek()'s use of .rindex_shown (vs frame_queue_peek_last()'s omission of it) should already handle that.

Also hasImage() and hasImageCurrent() were basically identical, so I merged them.