completely rewrite video-rendering using OpenGL 3 Core, and do YUV decoding in GL
Playing a random video I've been testing with now uses ~20% CPU rather than ~50%.

Some videos don't look correct - they have diagonal green stripes through them.  Not sure of the cause, but it seems to depend on the width (720, 716, 624 are problematic, whereas 540, 608, 640, 704, 1280, 1920 are not).  Usually just the colour is messed up, but I did find a case where the greyscale image was too.