LAVPMovie: clean up seeking code a bit, though it's still far from perfect
The major change here is the logic from -setPosition: (temporarily pausing, clipping, setting .busy) is now in -setCurrentTimeInMicroseconds:, so we finally have just one kind of seeking.  Alas, trying to step forward by 5 seconds still doesn't work, because seeking is to the nearest earlier keyframe (which can be before the time we started seeking *from*), but I'll fix that at some point.