eliminate the nasty sleep-looping in LAVPMovie's constructor
Its purpose was to get the first frame of the video ready to show before pausing.  But we can just re-use ->is_temporarily_unpaused_to_handle_seeking instead (technically this means we display nothing very briefly, but update as soon as the first frame is ready, but it's a very brief interval).