simplify decoder cleanup
stream_component_close()'s video/audio branches are now identical (thanks to Decoder became the owner of the FrameQueue and PacketQueue) so combine them, and merge into decoder_destroy().  Also merge in decoder_abort(), since it's not called elsewhere.  And then this is also the logical place to call packet_queue_destroy() and frame_queue_destory().