simplify video_refresh()
The remaining_time local, distinct from is->remaining_time was pretty confusing; he new code should have identical effect.  Also, the is->paused check was redundant with the one in refresh_loop_wait_event(), so remove it.  And finally, refresh_loop_wait_event() is now trivial, and terribly named (for historical reasons), so merge it in.  Not that video_refresh() is really an accurate name either, but it's closer.