complete rewrite video_refresh()
It seems to me the only thing it needs to do is skip any frames from before seeking (those where the serial is wrong) and then advance to the latest frame whose pts we have reached or exceeded.  So the new code just does that, completely dispensing with ->vidclk and calculating the delay and suchlike.  Possibly I'm overlooking something, but I think it's more likely that the old code just existed to accomodate using the video clock as master and similar, and perhaps was also complicated by running as part of the mainloop of ffplay.