change configure options used when compiling ffmpeg
Switch to --disable-network, because I don't want it.  Drop --enable-protocols in favour of using the defaults.  Add --disable-openssl because I shouldn't need it without network support.

Remove --enable-small because there's no obvious reason for it (so why not stick to the default).

Use --disable-programs instead of individual options to the same effect.

Add --disable-doc and --disable-sdl because we definitely don't need those (not sure what their defaults are).