stop limiting the size-in-bytes of the packet queues
It doesn't seem to make much difference in practice - the video packet queue sits at about 25 (i.e. MIN_FRAMES), and the audio one wanders from ~30 to ~100 (except after seeking, of course, when they're empty briefly).  And there's no obvious conceptual need for it either, that I can think of.  Perhaps it mattered in ffplay because of network streams, or something.

We're already limiting their length (in packets)