remove the frame-dropping code from video_thread()
Currently it seems to trigger once at the very start of playback (which seems iffy), and a few times if you do lots of seeks in a short period of time (e.g. if you hold down a seek-five-seconds key).  In both cases it should be fairly harmless to add the frame to the frame-queue and then skip it once it's in there.

For the actual case of frame-skipping when the CPU can't keep up with the movie, that's such a miserable experience as to be not worth bothering with, honestly.