make frame_queue_peek() return NULL if the queue is empty
i.e. put the length check inside it, rather than at the call sites.

I'm also eliminating decoder_get_current_frame_or_null() because the only caller is lavp_get_current_frame(), which is the only thing that modifies the front of the video frame queue, so the locking shouldn't be needed.