init the audio sample converter in audio_open() rather than audio_decode_frame()
The AVFrame format/channels/sample_rate fields are set in avcodec_decode_audio4() (which we call via avcodec_receive_frame()) by copying them from the AVCodecContext.  So it should be impossible for them to change mid-stream, and so we can just set up the converter once rather than potentially per-frame.