simplify audio clock updating
The new version generally seems to set the time about 40ms ahead of what the old code did, but possibly for the better (it's hard to tell with such small differences).  And the flaw in the new version (it's not accounting for buffered-but-not-yet-played audio) is probably there in the old code too, just buried under the complexity - it's very hard to tell though.