fix decoder_finished() sometimes returning true at the start of playback
Because the first serial we use is 0, and that's also what ->finished was (implicitly) initialised to.

I think this might be responsible for videos occasionally starting out displaying solid green, because the decoder_finished() call in read_thread() would mean we paused too early, before managing to have a first video frame ready.