use a single decoder thread, rather than one per stream
For the sake of code simplicity.  The only real drawback is that you can run out of video frames while waiting for frame-queue slots to decode an audio packet into (or vice versa), but that's easily handled by just making the frame queues longer.  The choice of 50 frames is arbitrary - it was enough to stop audio/video stuttering on the movies I tested it on.