don't flush the audio queue when pausing
It seems completely illogical to have been doing this, since it should just have meant we lose some audio for no reason.  Probably ~60ms worth, given the buffer sizes we're using.

This also removes the call to AudioQueuePrime(), but that looks like it should be pretty harmless since we're sending non-compressed audio to the queue.