copy audio directly from AVFrame to AudioQueueBufferRef rather than via ->audio_buf1
I was hoping for a small CPU win, but it appears to be non-existent.  It's still simpler though.  And it ought to work correctly with already-interleaved audio, which the old code clearly wouldn't have (it free'd the audio frame via decoder_advance_frame() and then went on to trying to read its data).