LAVPMovie: eliminate .lastPosition
The clock is never -1 any more so we don't need it.  We may occasionally get a slightly outdated clock, but that's probably better than getting what lastPosition was returning (since it's the *intended* target of seeking, but seeking is fuzzy, so actually it mean the clock ui could jump forward or backward a bit once the real after-seek clock time got set).