fix inadequate locking in decoder_peek_current_frame() and decoder_peek_current_frame_blocking()
If you pause and then seek immediately afterward you can end up with the audio/video threads and the decoders thread both calling these at the same time, and both deciding to call decoder_advance_frame() to skip a stale frames, potentially advancing when the queue is empty and leading to frameq_size going negative, after which everything is pretty broken.  This partly fixes it by holding the locks longer, but the remaining calls to decoder_advance_frame() are probably still sources of similar problems.

Prior to the switch to having seek-while-paused happen on the decoders thead this wasn't an issue because only the audio/video threads ever called decoder_advance_frame().