improve clock handling after opening and after seeking
There's no point starting the clock at -1, as the code in clock.m doesn't treat that specially, so it's pretty much the same as starting it at 0 (implicitly, since ObjC zeroes new objects).

And we should be trying to set the clock ASAP after seeking even if not paused, to avoid inadvertently skipping a video frame based on comparisons to the old clock.

The bonus fix of setting need_video_update_after_seeking_while_paused to true initially might also fix the bug where sometimes after opening a movie it show all green (because we failed to upload a new video frame).